Fix design flaw in geo::ChannelMapAlg interface.
The geo::ChannelMapAlg::SignalType() function has two
overloads, both virtual.  If a derived class overrides one,
but not the other, then the other signature is hidden.  This is
a developer error, but annoying.

The SignalType() virtual function has been replaced with two
non-virtual member functions, SignalTypeForChannel() and
SignalTypeForROPID(), one for each of the original two
overloads.

A derived class implementor may customize these functions by
providing overrides for SignalTypeForChannelImpl() and
SignalTypeForROPIDImpl(), which are called by the base class
implementations of SignalTypeForChannel() and
SignalTypeForROPID().