Fixing time synchronisation for PhotonBacktracker for ProtoDUNE HD MC.
I have observed the backtracker did not match OpHits to correct true particles and in general it failed to find corresponding true depositions. Looking at the OpFlash times w.r.t. the true times, the flash times came 250 us earlier. This fix is configuring the backtracker to account for that.